Code_App/Scripts/requirements.txt
PyGithub>=1.55
python-dotenv>=0.21
unidiff>=0.7
+requests>=2.28
transformers>=4.30
datasets>=2.12
